+++ src/components/CookieBanner.tsx
@@ -93,10 +93,10 @@ const CookieBanner = (): JSX.Element | null => {
                 </Heading>
                 <VStack spacing="2.5" alignItems="stretch">
                     <Text fontSize="sm">
-                        Owkin Connect being a young product, we use cookies to
-                        better understand your needs, expectations and pain
-                        points. By allowing us to use cookies you help us
-                        improve the quality and the accuracy of our services!
+                        Substra being a young product, we use cookies to better
+                        understand your needs, expectations and pain points. By
+                        allowing us to use cookies you help us improve the
+                        quality and the accuracy of our services!
                     </Text>
                     <Accordion allowMultiple allowToggle>
                         <AccordionItem border="none">
@@ -150,7 +150,7 @@ const CookieBanner = (): JSX.Element | null => {
                                     <Text fontSize="sm">
                                         Clarity is a free user behavior
                                         analytics tool that helps us understand
-                                        how users are interacting with Connect
+                                        how users are interacting with Substra
                                         through session replays and heatmaps.
                                     </Text>
                                 </Box>