+++ src/routes/login/Login.tsx
@@ -1,11 +1,10 @@
 import { useEffect } from 'react';
 
-import ReactPlayer from 'react-player';
 import { useLocation } from 'wouter';
 
 import {
-    AspectRatio,
     Box,
+    Flex,
     HStack,
     Text,
     VStack,
@@ -13,7 +12,8 @@ import {
     Button,
 } from '@chakra-ui/react';
 
-import OwkinLogoBlack from '@/assets/svg/owkin-logo-black.svg';
+import LoginBackground from '@/assets/login-background.png';
+import SubstraLogoBlack from '@/assets/svg/substra-full-name-logo.svg';
 import useAppDispatch from '@/hooks/useAppDispatch';
 import useAppSelector from '@/hooks/useAppSelector';
 import { useDocumentTitleEffect } from '@/hooks/useDocumentTitleEffect';
@@ -46,7 +46,7 @@ const Login = (): JSX.Element => {
     }, [forceLogout, authenticated, dispatch, setLocation, nextLocation]);
 
     return (
-        <HStack backgroundColor="#F7FAFC" flex="1">
+        <HStack backgroundColor="#F7FAFC" flex="1" spacing="0">
             <VStack
                 height="100vh"
                 width="500px"
@@ -55,44 +55,41 @@ const Login = (): JSX.Element => {
                 justifyContent="space-between"
             >
                 <Box padding="8">
-                    <OwkinLogoBlack />
+                    <SubstraLogoBlack />
                 </Box>
                 <Box alignItems="center" width="100%">
                     <LoginForm />
                 </Box>
                 <Text color="gray.500" fontSize="xs" paddingX="5" paddingY="8">
-                    © Owkin - All rights reserved. Frontend vers.
+                    Frontend vers.
                     {__APP_VERSION__}
                 </Text>
             </VStack>
-            <Box padding="12" flex="auto">
-                <AspectRatio ratio={16 / 9}>
-                    <ReactPlayer
-                        url="https://vimeo.com/541687596"
-                        controls
-                        width="100%"
-                        height="100%"
-                        config={{
-                            vimeo: {
-                                playerOptions: {
-                                    byline: true,
-                                    portrait: true,
-                                    title: true,
-                                },
-                            },
-                        }}
-                    />
-                </AspectRatio>
-                <Text marginY="4" fontSize="2xl">
+            <Flex
+                flexDirection="column"
+                height="100vh"
+                padding="10%"
+                backgroundImage={LoginBackground}
+                justifyContent="center"
+                style={{
+                    backgroundRepeat: 'no-repeat',
+                    backgroundSize: 'cover',
+                }}
+            >
+                <Text marginY="4" fontSize="2xl" color="white">
                     Unlock AI, break data silos, and protect privacy with our
                     Federated Learning software
                 </Text>
-                <Link href="https://owkin.com/owkin-connect/" isExternal>
-                    <Button variant="outline" colorScheme="black">
-                        <Text>Tell me more!</Text>
+                <Link
+                    href="https://owkin.com/owkin-connect/"
+                    isExternal
+                    color="white"
+                >
+                    <Button variant="outline" colorScheme="whiteAlpha">
+                        <Text color="white">Tell me more!</Text>
                     </Button>
                 </Link>
-            </Box>
+            </Flex>
         </HStack>
     );
 };