+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Increase page size to 30
 -   CP Workflow graph: Allow a higher min zoom to show more tasks
 -   CP Workflow graph: update layout for predict tasks and test tasks
+-   Change Connect branding to Substra
 
 ### Fixed